Refactor LLM executors and document core APIs
