+++ llm/core/model_runner.py
@@ -1,0 +1,28 @@
+# Copyright (c) PyPTO Contributors.
+# This program is free software, you can redistribute it and/or modify it under the terms and conditions of
+# CANN Open Software License Agreement Version 2.0 (the "License").
+# Please refer to the License for details. You may not use this file except in compliance with the License.
+# THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
+# INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
+# See LICENSE in the root of the software repository for the full text of the License.
+# -----------------------------------------------------------------------------------------------------------
+
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+
+from .types import DecodeBatch, DecodeResult, PrefillBatch, PrefillResult, RuntimeModel
+
+
+class ModelRunner(ABC):
+    """Runtime interface for compiled kernels registered to one model."""
+
+    @abstractmethod
+    def run_prefill(self, model: RuntimeModel, batch: PrefillBatch) -> PrefillResult:
+        """Run the compiled prefill path for one batch."""
+        raise NotImplementedError
+
+    @abstractmethod
+    def run_decode(self, model: RuntimeModel, batch: DecodeBatch) -> DecodeResult:
+        """Run the compiled decode path for one batch."""
+        raise NotImplementedError
